feat:抽离sendPV API
