+++ src/utils/index.ts
@@ -23,3 +23,18 @@ export function sendAESEvent(eventName: string, ...customArgs: any[]): void {
     setTimeout(() => sendLog(), 0);
   });
 }
+
+export function sendPV(location): void {
+  const pageId = location.pathname.replace(/^\/developer-site/, '');
+  if (window.AES && window.AES.getConfig('page_id') !== pageId) {
+    // 如果之前有pageId 则先发送leave埋点
+    if (window.AES.getConfig('page_id')) {
+      window.AESPluginPV && window.AESPluginPV.sendLeave();
+    }
+    // 修改PageId并发送PV埋点
+    window.AES.setConfig({
+      page_id: pageId,
+    });
+    window.AESPluginPV && window.AESPluginPV.sendPV();
+  }
+};
